Updated the website for once
Had to recreate the website as I lost the login for the old one. Had to change the backend. Updated the version of php and wordpress. The website is still a bit of a meme, but there is more that is going to be put into it now that I have a working backend.